+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
   - Prevent submit on empty input
+  - Handle search error, refs #6
 
 ## [2020.12.10]
   - Initial UI release